+++ CHANGELOG.md
@@ -3,10 +3,26 @@
 All notable changes for each version of this project will be documented in this file.
 
 ## 14.1.0
-- Added ability to auto-size columns to the size of their cells and header content on initialization by setting width `auto`:
-```
-<column width='auto' ...>
-```
+
+### New Features
+- `IgxCombo` and  `IgxSimpleComboComponent`
+    - `filterFunction` input is added. The new property allows changing of the way filtering is done in the combos. By default filtering is made over the values in combo's data when it is a collection of primitive values, or over the values as defined in `displayKey` of the combo. If custom filtering function is provided filtering will be done as specified in the provided function.
+    - `filteringOptions` are extended and now contains `filterable` and `filteringKey` properties. Setting `filterable` determines whether combo will be filterable. By default filtering is done over the data value when they are primitive, or over the field of the values equal to `displayKey`. `filteringKey` allows to filter data by any data related key.
+    - **Breaking Changes** - `filterable` property of `IgxComboComponent` is now deprecated and will be removed in future version. Use `filteringOptions.filterable` instead.
+
+- `igxPivotGrid`
+    - Add option to template the pivot value chip content:
+    ```
+    <ng-template igxPivotValueChip let-value>
+            {{ value.member }}
+    </ng-template>
+    ``` 
+
+- `igxGrid`
+	- Added ability to auto-size columns to the size of their cells and header content on initialization by setting width `auto`:
+	```
+	<column width='auto' ...>
+	```
 
 
 
@@ -18,7 +34,8 @@ All notable changes for each version of this project will be documented in this 
     - The `label` attribute has been changed to `aria-labelledby` and can be set by a latterly added input property `labelledBy`.
 
 ### New Features
-
+- `IgxCombo` and  `IgxSimpleComboComponent`
+    - `filterFunction` input is added. The new property allows changing of the way filtering is done in the combos. By default filtering is made over the values in combo's data when it is a collection of primitive values, or over the values as defined in `displayKey` of the combo. If custom filtering function is provided filtering will be done as specified in the provided function.
 
 ### General
 - Updating dependency to Angular 14